+++ CHANGELOG.md
@@ -5,8 +5,7 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
-
-## **WORK IN PROGRESS**
+## 1.3.0 (2022-11-04)
 
 * (theimo1221) Add persistence of device settings
 * (theimo1221) Update packages